many fixes and improvements

+++ packages/server/scripts/deleteUsers.ts
@@ -1,7 +1,15 @@
-// delete records in the users table
 import { PrismaClient } from "@prisma/client";
+import { confirm } from "./confirm";
+
 const prisma = new PrismaClient();
+
 async function main() {
+  const c = await confirm("Delete all users");
+  if (!c) {
+    console.log("Aborting");
+    return;
+  }
+
   await prisma.user.deleteMany({});
 }
 main()
